Refactor file and variable naming
