--- lib/net/scg_gcn.py
@@ -1,129 +1,0 @@
-import torch
-import torch.nn as nn
-
-
-class SCG_block(nn.Module):
-    def __init__(self, in_ch, hidden_ch=6, node_size=(32,32), add_diag=True, dropout=0.2):
-        super(SCG_block, self).__init__()
-        self.node_size = node_size
-        self.hidden = hidden_ch
-        self.nodes = node_size[0]*node_size[1]
-        self.add_diag = add_diag
-        self.pool = nn.AdaptiveAvgPool2d(node_size)
-
-        self.mu = nn.Sequential(
-            nn.Conv2d(in_ch, hidden_ch, 3, padding=1, bias=True),
-            nn.Dropout(dropout),
-        )
-
-        self.logvar = nn.Sequential(
-            nn.Conv2d(in_ch, hidden_ch, 1, 1, bias=True),
-            nn.Dropout(dropout),
-        )
-
-    def forward(self, x):
-        B, C, H, W = x.size()
-        gx = self.pool(x)
-
-        mu, log_var = self.mu(gx), self.logvar(gx)
-
-        if self.training:
-            std = torch.exp(log_var.reshape(B, self.nodes, self.hidden))
-            eps = torch.randn_like(std)
-            z = mu.reshape(B, self.nodes, self.hidden) + std*eps
-        else:
-            z = mu.reshape(B, self.nodes, self.hidden)
-
-        A = torch.matmul(z, z.permute(0, 2, 1))
-        A = torch.relu(A)
-
-        Ad = torch.diagonal(A, dim1=1, dim2=2)
-        mean = torch.mean(Ad, dim=1)
-        gama = torch.sqrt(1 + 1.0 / mean).unsqueeze(-1).unsqueeze(-1)
-
-        dl_loss = gama.mean() * torch.log(Ad[Ad<1]+ 1.e-7).sum() / (A.size(0) * A.size(1) * A.size(2))
-
-        kl_loss = -0.5 / self.nodes * torch.mean(
-            torch.sum(1 + 2 * log_var - mu.pow(2) - log_var.exp().pow(2), 1)
-        )
-
-        loss = kl_loss - dl_loss
-
-        if self.add_diag:
-            diag = []
-            for i in range(Ad.shape[0]):
-                diag.append(torch.diag(Ad[i, :]).unsqueeze(0))
-
-            A = A + gama * torch.cat(diag, 0)
-            # A = A + A * (gama * torch.eye(A.size(-1), device=A.device).unsqueeze(0))
-
-        # A = laplacian_matrix(A, self_loop=True)
-        A = self.laplacian_matrix(A, self_loop=True)
-        # A = laplacian_batch(A.unsqueeze(3), True).squeeze()
-
-        z_hat = gama.mean() * \
-                mu.reshape(B, self.nodes, self.hidden) * \
-                (1. - log_var.reshape(B, self.nodes, self.hidden))
-
-        return A, gx, loss, z_hat
-
-    @classmethod
-    def laplacian_matrix(cls, A, self_loop=False):
-        '''
-        Computes normalized Laplacian matrix: A (B, N, N)
-        '''
-        if self_loop:
-            A = A + torch.eye(A.size(1), device=A.device).unsqueeze(0)
-        # deg_inv_sqrt = (A + 1e-5).sum(dim=1).clamp(min=0.001).pow(-0.5)
-        deg_inv_sqrt = (torch.sum(A, 1) + 1e-5).pow(-0.5)
-
-        LA = deg_inv_sqrt.unsqueeze(-1) * A * deg_inv_sqrt.unsqueeze(-2)
-
-        return LA
-
-
-class GCN_Layer(nn.Module):
-    def __init__(self, in_features, out_features, bnorm=True,
-                 activation=nn.ReLU(), dropout=None):
-        super(GCN_Layer, self).__init__()
-        self.bnorm = bnorm
-        fc = [nn.Linear(in_features, out_features)]
-        if bnorm:
-            fc.append(BatchNorm_GCN(out_features))
-        if activation is not None:
-            fc.append(activation)
-        if dropout is not None:
-            fc.append(nn.Dropout(dropout))
-        self.fc = nn.Sequential(*fc)
-
-    def forward(self, data):
-        x, A = data
-        y = self.fc(torch.bmm(A, x))
-
-        return [y, A]
-
-
-def weight_xavier_init(*models):
-    for model in models:
-        for module in model.modules():
-            if isinstance(module, nn.Conv2d) or isinstance(module, nn.Linear):
-                # nn.init.xavier_normal_(module.weight)
-                nn.init.orthogonal_(module.weight)
-                # nn.init.kaiming_normal_(module.weight)
-                if module.bias is not None:
-                    module.bias.data.zero_()
-            elif isinstance(module, nn.BatchNorm2d):
-                module.weight.data.fill_(1)
-                module.bias.data.zero_()
-
-
-class BatchNorm_GCN(nn.BatchNorm1d):
-    '''Batch normalization over GCN features'''
-
-    def __init__(self, num_features):
-        super(BatchNorm_GCN, self).__init__(num_features)
-
-    def forward(self, x):
-        return super(BatchNorm_GCN, self).forward(x.permute(0, 2, 1)).permute(0, 2, 1)
-
-
