+++ train.py
@@ -1,0 +1,1 @@
+# TODO: implement a similar system to that of ultralytic's yolov5 or other ML CLIs
